feat(broker):Added pipe socket conditions for different systems
