+++ broker/broker.go
@@ -5,7 +5,6 @@ import (
 	encJson "encoding/json"
 	"errors"
 	"fmt"
-	"github.com/natefinch/npipe"
 	"net"
 	"net/http"
 	"os"
@@ -304,57 +303,6 @@ func (b *Broker) StartUnixSocketClientListening(socketPath string, unixSocket bo
 	tmpDelay := 10 * ACCEPT_MIN_SLEEP
 	for {
 		conn, err := l.Accept()
-		if err != nil {
-			if ne, ok := err.(net.Error); ok && ne.Temporary() {
-				log.Error(
-					"Temporary Client Accept Error(%v), sleeping %dms",
-					zap.Error(ne),
-					zap.Duration("sleeping", tmpDelay/time.Millisecond),
-				)
-
-				time.Sleep(tmpDelay)
-				tmpDelay *= 2
-				if tmpDelay > ACCEPT_MAX_SLEEP {
-					tmpDelay = ACCEPT_MAX_SLEEP
-				}
-			} else {
-				log.Error("Accept error", zap.Error(err))
-			}
-			continue
-		}
-
-		tmpDelay = ACCEPT_MIN_SLEEP
-		go func() {
-			err := b.handleConnection(CLIENT, conn)
-			if err != nil {
-				conn.Close()
-			}
-		}()
-	}
-}
-
-// StartPipeSocketListening We use the open source npipe library to support pipe communication in windows
-func (b *Broker) StartPipeSocketListening(pipeName string, usePipe bool) {
-	var err error
-	var ln *npipe.PipeListener
-
-	for {
-		if usePipe {
-			fmt.Println(pipeName)
-			ln, err = npipe.Listen(pipeName)
-			log.Info("Start Listening client on ", zap.String("pipeName", pipeName))
-		}
-		if err == nil {
-			break // successfully listening
-		}
-		log.Error("Error listening on ", zap.Error(err))
-		time.Sleep(1 * time.Second)
-	}
-
-	tmpDelay := 10 * ACCEPT_MIN_SLEEP
-
-	for {
-		conn, err := ln.Accept()
 		if err != nil {
 			if ne, ok := err.(net.Error); ok && ne.Temporary() {
 				log.Error(
